Major Overhaul of all nodes. BREAKING CHANGES
Converted all nodes to use generated code (see gen_scripts).
All published nodes can be generated directly by running `nodejs run_build.js` and copying the desired files from the build dir to the module root.  The exception is IoTData which requires patching with the diffs in gen_scripts.

Under the new scheme ALL api functionality is exposed for supported services.   This means that the parameters for each are currently missing, all parameters must be specified in the message object..